dom/tests/mochitest/geolocation/test_handlerSpinsEventLoop.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=911595
-->
<head>
  <title>Test for spinning the event loop inside position handlers</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="geolocation_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=911595 ">Mozilla Bug 911595</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/*
 * In bug 911595 , spinning the event loop from inside position
 * handlers could cause both success and error callbacks to be
 * fired for the same request if that request has a small timeout.
 */

SimpleTest.waitForExplicitFinish();
SimpleTest.requestFlakyTimeout("untriaged");

resume_geolocationProvider(function() {
  force_prompt(true, test1);
});

function spinEventLoopAndSetTimeout() {
  if (successCallbackCalled || errorCallbackCalled) {
    // this should only be called once from either callback
    return;
  }

  SpecialPowers.spinEventLoop(window);

  setTimeout(function() {
    ok(successCallbackCalled != errorCallbackCalled, "Ensure only one callback is called");
    SimpleTest.finish();
  }, 5);
}

var successCallbackCalled = false;
function successCallback(position) {
  spinEventLoopAndSetTimeout();
  successCallbackCalled = true;
}

var errorCallbackCalled = false;
function errorCallback(error) {
  spinEventLoopAndSetTimeout();
  errorCallbackCalled = true;
}

function test1() {
  navigator.geolocation.getCurrentPosition(successCallback, errorCallback, {timeout: 1});
}
</script>
</pre>
</body>
</html>
